Match g_data hp setter, persona init, and mark existing bodies

+++ include/Main/g_data.h
@@ -280,7 +280,7 @@ void datSetFlag(s32 bit, u8 enabled);
 void datClearFlagAll();
 u16 datGetEquipmentId(s16 pcId, u16 equipmentIdx);
 u8 datGetEquipmentEffect(s16 pcId, u16 equipmentIdx);
-void datInitPersona(u32 pcId);
+void datInitPersona(s16 pcId);
 void datCompendiumInit();
 s16 datGetActiveSocialLink();
 s8 datGetSocialLinkLevel(s16 socialLink);
